Put modules somewhere, create CommandTable
