--- ndn_hydra/repo/global_view/__init__.py
@@ -1,10 +1,0 @@
-# -------------------------------------------------------------
-#  @Project: NDN Hydra
-#  @Date:    2021-01-25
-#  @Authors: Please check AUTHORS.rst
-#  @Source-Code:   https://github.com/justincpresley/ndn-hydra
-#  @Documentation: https://ndn-hydra.readthedocs.io
-#  @Pip-Library:   https://pypi.org/project/ndn-hydra
-# -------------------------------------------------------------
-
-from .global_view import GlobalView
