+++ ndn_hydra/repo/group_messages/claim.py
@@ -13,7 +13,7 @@ import copy
 from typing import Callable
 import time
 from ndn.encoding import *
-from ndn_hydra.repo.global_view.global_view import GlobalView
+from ndn_hydra.repo.modules.global_view import GlobalView
 from ndn_hydra.repo.group_messages.specific_message import SpecificMessage
 
 class ClaimTypes:
