+++ ndn_hydra/repo/group_messages/expire.py
@@ -13,7 +13,7 @@ from typing import Callable
 from ndn.encoding import *
 import time
 from ndn_hydra.repo.group_messages.specific_message import SpecificMessage
-from ndn_hydra.repo.global_view.global_view import GlobalView
+from ndn_hydra.repo.modules.global_view import GlobalView
 
 class ExpireMessageTypes:
     SESSION_ID = 83
