+++ ndn_hydra/repo/group_messages/store.py
@@ -12,7 +12,7 @@
 from typing import Callable
 from ndn.encoding import *
 from ndn_hydra.repo.group_messages.specific_message import SpecificMessage
-from ndn_hydra.repo.global_view.global_view import GlobalView
+from ndn_hydra.repo.modules.global_view import GlobalView
 
 class StoreMessageTypes:
     SESSION_ID = 83
