+++ ndn_hydra/repo/handles/query_handle.py
@@ -14,7 +14,7 @@ import logging
 from ndn.app import NDNApp
 from ndn.encoding import Name, ContentType, Component
 from ndn.storage import Storage
-from ndn_hydra.repo.global_view.global_view import GlobalView
+from ndn_hydra.repo.modules.global_view import GlobalView
 from ndn_hydra.repo.protocol.base_models import FileList, File
 
 class QueryHandle(object):
